fix: improve dynamic tool discovery for not development tasks
