--- src/mcp/tools/simulator-environment/__tests__/index.test.ts
@@ -1,90 +1,0 @@
-/**
- * Tests for simulator-environment workflow metadata
- */
-import { describe, it, expect } from 'vitest';
-import { workflow } from '../index.js';
-
-describe('simulator-environment workflow metadata', () => {
-  describe('Workflow Structure', () => {
-    it('should export workflow object with required properties', () => {
-      expect(workflow).toHaveProperty('name');
-      expect(workflow).toHaveProperty('description');
-      expect(workflow).toHaveProperty('platforms');
-      expect(workflow).toHaveProperty('targets');
-      expect(workflow).toHaveProperty('projectTypes');
-      expect(workflow).toHaveProperty('capabilities');
-    });
-
-    it('should have correct workflow name', () => {
-      expect(workflow.name).toBe('Simulator Environment Configuration');
-    });
-
-    it('should have correct description', () => {
-      expect(workflow.description).toBe(
-        'Tools for configuring iOS Simulator environment settings including appearance, location services, and network conditions. Perfect for testing apps under various environmental conditions.',
-      );
-    });
-
-    it('should have correct platforms array', () => {
-      expect(workflow.platforms).toEqual(['iOS']);
-    });
-
-    it('should have correct targets array', () => {
-      expect(workflow.targets).toEqual(['simulator']);
-    });
-
-    it('should have correct projectTypes array', () => {
-      expect(workflow.projectTypes).toEqual(['project', 'workspace']);
-    });
-
-    it('should have correct capabilities array', () => {
-      expect(workflow.capabilities).toEqual([
-        'environment-config',
-        'appearance',
-        'location',
-        'network-simulation',
-      ]);
-    });
-  });
-
-  describe('Workflow Validation', () => {
-    it('should have valid string properties', () => {
-      expect(typeof workflow.name).toBe('string');
-      expect(typeof workflow.description).toBe('string');
-      expect(workflow.name.length).toBeGreaterThan(0);
-      expect(workflow.description.length).toBeGreaterThan(0);
-    });
-
-    it('should have valid array properties', () => {
-      expect(Array.isArray(workflow.platforms)).toBe(true);
-      expect(Array.isArray(workflow.targets)).toBe(true);
-      expect(Array.isArray(workflow.projectTypes)).toBe(true);
-      expect(Array.isArray(workflow.capabilities)).toBe(true);
-
-      expect(workflow.platforms.length).toBeGreaterThan(0);
-      expect(workflow.targets.length).toBeGreaterThan(0);
-      expect(workflow.projectTypes.length).toBeGreaterThan(0);
-      expect(workflow.capabilities.length).toBeGreaterThan(0);
-    });
-
-    it('should contain expected platform values', () => {
-      expect(workflow.platforms).toContain('iOS');
-    });
-
-    it('should contain expected target values', () => {
-      expect(workflow.targets).toContain('simulator');
-    });
-
-    it('should contain expected project type values', () => {
-      expect(workflow.projectTypes).toContain('project');
-      expect(workflow.projectTypes).toContain('workspace');
-    });
-
-    it('should contain expected capability values', () => {
-      expect(workflow.capabilities).toContain('environment-config');
-      expect(workflow.capabilities).toContain('appearance');
-      expect(workflow.capabilities).toContain('location');
-      expect(workflow.capabilities).toContain('network-simulation');
-    });
-  });
-});
