+++ src/mcp/tools/simulator-management/__tests__/set_sim_appearance.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect } from 'vitest';
 import { z } from 'zod';
 import setSimAppearancePlugin, { set_sim_appearanceLogic } from '../set_sim_appearance.ts';
-import { createMockExecutor, createMockFileSystemExecutor } from '../../../../utils/command.js';
+import { createMockExecutor } from '../../../../utils/command.js';
 
 describe('set_sim_appearance plugin', () => {
   describe('Export Field Validation (Literal)', () => {
