+++ src/mcp/tools/simulator-management/__tests__/set_simulator_location.test.ts
