+++ src/mcp/tools/simulator-management/boot_sim.ts
@@ -1,0 +1,2 @@
+// Re-export from simulator-workspace to avoid duplication
+export { default } from '../simulator-shared/boot_sim.ts';
