+++ src/mcp/tools/simulator-management/list_sims.ts
@@ -1,0 +1,2 @@
+// Re-export from simulator-shared to avoid duplication
+export { default } from '../simulator-shared/list_sims.ts';
