+++ src/mcp/tools/simulator-management/open_sim.ts
@@ -1,0 +1,2 @@
+// Re-export from simulator-shared to avoid duplication
+export { default } from '../simulator-shared/open_sim.ts';
