+++ src/mcp/tools/simulator-management/reset_simulator_location.ts
@@ -1,8 +1,8 @@
 import { z } from 'zod';
-import { ToolResponse } from '../../../types/common.js';
-import { log } from '../../../utils/index.js';
-import { CommandExecutor, getDefaultCommandExecutor } from '../../../utils/index.js';
-import { createTypedTool } from '../../../utils/typed-tool-factory.js';
+import { ToolResponse } from '../../../types/common.ts';
+import { log } from '../../../utils/index.ts';
+import { CommandExecutor, getDefaultCommandExecutor } from '../../../utils/index.ts';
+import { createTypedTool } from '../../../utils/typed-tool-factory.ts';
 
 // Define schema as ZodObject
 const resetSimulatorLocationSchema = z.object({
